Removed the custom pending transition
Hey,

good library thanks!

I want to use the default android pending transition in my apps, but since there is already a custom transition defined in the library, i have to override it in every app. In my opinion we should remove the custom transition in the library and let the developers decide to use a custom animation (if they want).